refactor: the workflow is no longer an agent
